requirements.txt
flatbuffers==2.0
numpy==1.21.2
onnx==1.10.1
onnxruntime==1.8.1
Pillow==8.3.1
protobuf==3.17.3
six==1.16.0
torch==1.9.0
torchextractor==0.3.0
torchvision==0.10.0
typing-extensions==3.10.0.0
